Define the knex constant in repository.js
